Modified  CLOptionText and CLOptionPresent to accept the command line arguments as a tuple and process accordingly. This allows for multiple arguments to represent the same command
Modified the help text output. Now reads the values stored in CLOptions so if we need to change the long or short command, the help text will reflect the change
Also changed the help text format to be more readable and represent the possible values

moved the detection of the --help command line option to occur first. This ensures you can call help halfway through typing a command to see the help text