add "name=" modifier to textfields, checkboxes and dropdowns.
This means you can specify an alternate name for the resulting output. e.g. `--textfield "Some Long Text",name=example1` will result in output where the value of the textfield will be returned as `example1 : <value>` or in json `{"example1" : "<value>"}`

for checkboxes it will be `--checkbox "Some Long Text",name=example2`

for dropdowns it will be `--selecttitle "Some Long Text",name=example3`

In addition, the processing for checkboxes has been cleaned up to be consistent with how textfields and dropdowns are handled.