addressing #476
This is caused by a change in 2.5.4 where I thought I was fixing the behaviour where running swiftdialog with no options launches a boilerplate dialog. What I _want_ to have happen is nothing it launched but doing this at the location I was doing it causes notification processing to exit prematurely and thus actions were not being processed.