Added buttonsize as a command line and control file variable.
Added support for mini, small and large button button sizes, using the build in macOS sizes. The button size will default to the current value of regular if not specified or not valid.

Added support for button size to the builder for testing and JSON generation.

Tested with control files, json input and command line, and with incorrect values.

Button style is applied to button1, button2 and the info button. It is also applied iin the normal and stacked button layout.

 It does not scale the help button, which is a fixed size.